+++ .eslintrc.js
@@ -4,13 +4,24 @@ module.exports = {
     es2021: true,
     node: true,
   },
-  extends: ["eslint:recommended", "plugin:import/recommended", "plugin:@typescript-eslint/recommended"],
+  extends: [
+    "eslint:recommended",
+    "plugin:import/recommended",
+    "plugin:@typescript-eslint/recommended"
+  ],
   parserOptions: {
     ecmaVersion: 2021,
     sourceType: "module",
   },
   parser: "@typescript-eslint/parser",
-  plugins: ["import", "node", "promise", "unused-imports", "@typescript-eslint"],
+  plugins: [
+    "import",
+    "node",
+    "promise",
+    "unused-imports",
+    "@typescript-eslint",
+    "jsdoc" // Add the jsdoc plugin
+  ],
   rules: {
     indent: ["error", 2, { MemberExpression: 1, SwitchCase: 1 }],
     "linebreak-style": ["error", "unix"],
@@ -23,10 +34,28 @@ module.exports = {
       { alphabetize: { caseInsensitive: true, order: "asc" } },
     ],
     "import/newline-after-import": "error",
+    "@typescript-eslint/no-explicit-any": "warn",
+
+    // JSDoc rules for TypeScript files
+    "jsdoc/require-jsdoc": ["warn", {
+      "contexts": [
+        "TSInterfaceDeclaration",
+        "TSMethodSignature",
+        "TSFunctionType",
+        "TSConstructorType"
+      ]
+    }],
+    "jsdoc/require-param": "error",
+    "jsdoc/require-param-description": "error",
+    "jsdoc/require-returns": "error",
+    "jsdoc/require-description": "error"
   },
   settings: {
     "import/resolver": {
-      "typescript": {}
+      typescript: {},
+      node: {
+        extensions: [".js", ".jsx", ".ts", ".tsx"]
+      }
     }
   }
 };