+++ package/config/provider.config.ts
@@ -8,22 +8,23 @@
  */
 import * as fs from "fs";
 import { constant } from "../constants/server.constant";
-import ConfigType from "./types.config";
+import { ApplicationContext } from "./../context/application.context";
+import GenericObject from "./types.config";
 
-export const configProvider = (): ConfigType => {
+export const configProvider = (): GenericObject => {
   // eslint-disable-next-line no-useless-catch
   try {
     let configFilePath: string =
       process.env[constant.entityStatus.WRAPPID_SERVICE_CONFIG_PATH];
-    let wrappidConfig: ConfigType;
+    let wrappidConfig: any;
     if (!configFilePath) {
       configFilePath = "./config.json";
     }
     // eslint-disable-next-line prefer-const
     wrappidConfig = JSON.parse(fs.readFileSync(configFilePath, "utf-8"));
-    console.log("###########################################");
-    console.log("configProvider found");
-    console.log("###########################################");
+    Object.keys(wrappidConfig).forEach((key: string) => {
+      ApplicationContext.setContext(key, wrappidConfig[key]);
+    });
     return wrappidConfig;
   } catch (error: any) {
     throw error;