+++ package/context/application.context.ts
@@ -1,0 +1,16 @@
+type GenericObject = {
+  [key: string]: any;
+};
+
+export class ApplicationContext {
+  private static context: GenericObject = {};
+  
+  static getContext(key: string): GenericObject {
+    return ApplicationContext.context[key];
+  }
+  
+  static setContext(key: string, value: any): void {
+    ApplicationContext.context[key] = value;
+  }
+}
+