+++ package/index.ts
@@ -1,8 +1,9 @@
-import { app } from "./app";
+import { App } from "./app";
 import { cacheActions } from "./cache/cache.actions";
 import { communicate } from "./communication/communicate.communicator";
 import { configProvider } from "./config/provider.config";
 import { constant, DEFAULT_PORT } from "./constants/server.constant";
+import { ApplicationContext } from "./context/application.context";
 import { databaseActions } from "./database/actions.database";
 import {
   getDatabases,
@@ -10,6 +11,7 @@ import {
   getColumns,
 } from "./database/helper.database";
 import { databaseProvider } from "./database/provider.database";
+import { paymentActions } from "./payment/payment.action";
 import {
   ControllersRegistry,
   FunctionsRegistry,
@@ -19,10 +21,12 @@ import {
   ValidationsRegistry,
   RoutesRegistry
 } from "./registry/index";
-
+import { APIService } from "./service/api.service";
 
 
 export {
+  APIService,
+  ApplicationContext,
   configProvider,
   /**
    * Registry
@@ -47,6 +51,7 @@ export {
   databaseProvider,
   databaseActions,
   cacheActions,
+  paymentActions,
   getDatabases,
   getTables,
   getColumns,
@@ -59,5 +64,8 @@ export {
   /**
    * 
    */
-  app
+  App,
+  /**
+   * 
+   */
 };