+++ CHANGELOG.md
@@ -2,6 +2,149 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.74](https://github.com/wrappid/service-core/compare/v0.0.73...v0.0.74) (2024-05-01)
+
+
+### Bug Fixes
+
+* **core:** :zap: added ThemesSchemas to data specific functions ([489d7b7](https://github.com/wrappid/service-core/commit/489d7b7df5db51a33249d12e494ea3f540335689)), closes [#128](https://github.com/wrappid/service-core/issues/128)
+* **global:** :bug: data sync api fix ([8ec977c](https://github.com/wrappid/service-core/commit/8ec977cf0233ddebd704cfab092a7b7bfaa88a17)), closes [#128](https://github.com/wrappid/service-core/issues/128)
+
+### [0.0.73](https://github.com/wrappid/service-core/compare/v0.0.72...v0.0.73) (2024-04-27)
+
+
+### Bug Fixes
+
+* **config:** :package: add missing package axios ([5998424](https://github.com/wrappid/service-core/commit/59984242400d182d8a9a89126a3a66b3d9446e78)), closes [#142](https://github.com/wrappid/service-core/issues/142)
+
+### [0.0.72](https://github.com/wrappid/service-core/compare/v0.0.71...v0.0.72) (2024-04-26)
+
+### [0.0.71](https://github.com/wrappid/service-core/compare/v0.0.70...v0.0.71) (2024-04-26)
+
+
+### Features
+
+* **global:** :sparkles: master data table clone api added ([2895e85](https://github.com/wrappid/service-core/commit/2895e850b76232e69c17cf262a13cbd8aca4edc1)), closes [#138](https://github.com/wrappid/service-core/issues/138)
+
+
+### Bug Fixes
+
+* **global:** :ambulance: whereclause authRequired  fix ([bb9265d](https://github.com/wrappid/service-core/commit/bb9265d5deda74b2e2ea027bc29baec2b96a29b5)), closes [#138](https://github.com/wrappid/service-core/issues/138)
+
+### [0.0.70](https://github.com/wrappid/service-core/compare/v0.0.69...v0.0.70) (2024-04-26)
+
+
+### Features
+
+* **global:** :bug: fix constant for registry ([f022297](https://github.com/wrappid/service-core/commit/f022297e9082cdf35976f739ae53f93ee669f0ff)), closes [#129](https://github.com/wrappid/service-core/issues/129)
+* **global:** :sparkles: get registry data fix ([f916dfd](https://github.com/wrappid/service-core/commit/f916dfd2bab4682c74cc23d1973b5408d231ec39)), closes [#129](https://github.com/wrappid/service-core/issues/129)
+* **global:** :sparkles: registry content read in application context ([788d6c6](https://github.com/wrappid/service-core/commit/788d6c67fb85a6d5f585d26a29ae46aab7074eb1)), closes [#129](https://github.com/wrappid/service-core/issues/129)
+
+
+### Bug Fixes
+
+* **global:** :ambulance: fix yup validation import ([f7dee33](https://github.com/wrappid/service-core/commit/f7dee33e4edd7dcbd2bfe4d6ddc47d7d89eb2da1)), closes [#129](https://github.com/wrappid/service-core/issues/129)
+* **global:** :ambulance: set routestRegistry  fix ([145ac7d](https://github.com/wrappid/service-core/commit/145ac7d0ced0660e0e052eb9f050e786ec35a8a5)), closes [#129](https://github.com/wrappid/service-core/issues/129)
+* **global:** :bug: putDatabaseModelFunc databaseprobider model rawattribute fix ([6fab725](https://github.com/wrappid/service-core/commit/6fab725e0668fe855ee973efd4e591a29cded986)), closes [#129](https://github.com/wrappid/service-core/issues/129)
+
+### [0.0.69](https://github.com/wrappid/service-core/compare/v0.0.68...v0.0.69) (2024-04-26)
+
+### [0.0.68](https://github.com/wrappid/service-core/compare/v0.0.67...v0.0.68) (2024-04-23)
+
+
+### Features
+
+* **global:** :recycle: update code from development ([91a29bd](https://github.com/wrappid/service-core/commit/91a29bd6f01755393115a5ce6d31e8d145e0de06)), closes [#132](https://github.com/wrappid/service-core/issues/132)
+* **global:** :sparkles: get registry file data api added ([da3d030](https://github.com/wrappid/service-core/commit/da3d030cc19d61148e22d015bce09343c1d8d3ed)), closes [#132](https://github.com/wrappid/service-core/issues/132)
+* **global:** :zap: database call for routes improve ([e9693ad](https://github.com/wrappid/service-core/commit/e9693ad0c9f82b66b5f7f4f7bd97e92445a0d876)), closes [#132](https://github.com/wrappid/service-core/issues/132)
+
+
+### Bug Fixes
+
+* **global:** :fire: api caller ([ec0c37e](https://github.com/wrappid/service-core/commit/ec0c37ed49895fb9c8415ad7f6a71c07ba5932ad)), closes [#129](https://github.com/wrappid/service-core/issues/129)
+* **global:** :fire: fix object type ([368d7fb](https://github.com/wrappid/service-core/commit/368d7fbff3a31226062d79269922bc697c302932)), closes [#132](https://github.com/wrappid/service-core/issues/132)
+* **global:** :recycle: code lint fix ([7795011](https://github.com/wrappid/service-core/commit/779501148729c18cc6623f547a137cd8a85bbff1)), closes [#132](https://github.com/wrappid/service-core/issues/132)
+
+### [0.0.67](https://github.com/wrappid/service-core/compare/v0.0.66...v0.0.67) (2024-04-20)
+
+
+### Features
+
+* **global:** :bug: delivery oriented fix ([a7ab924](https://github.com/wrappid/service-core/commit/a7ab924de3bf324953cec4ffa257552dd8d25400)), closes [#128](https://github.com/wrappid/service-core/issues/128)
+* **global:** :sparkles: create API data/sync/:model ([1460d02](https://github.com/wrappid/service-core/commit/1460d024afe68e93a43785a1932a025a47e90c71)), closes [#128](https://github.com/wrappid/service-core/issues/128)
+
+
+### Bug Fixes
+
+* **core:** :zap: fixing the issue related to data api(s) ([f5ebcc7](https://github.com/wrappid/service-core/commit/f5ebcc78210abcefe5de2db61fa725d50b4a6d2a)), closes [#128](https://github.com/wrappid/service-core/issues/128)
+* **global:** :bug: data sync API fix multiple data insert ([db13e8b](https://github.com/wrappid/service-core/commit/db13e8b699d688437479192eaae795e931c40e57)), closes [#128](https://github.com/wrappid/service-core/issues/128)
+
+### [0.0.66](https://github.com/wrappid/service-core/compare/v0.0.65...v0.0.66) (2024-04-18)
+
+
+### Bug Fixes
+
+* **global:** :bug: auther key change ([8194863](https://github.com/wrappid/service-core/commit/81948630261ae507263c1565ab19552591aa119e)), closes [#25](https://github.com/wrappid/service-core/issues/25)
+
+### [0.0.65](https://github.com/wrappid/service-core/compare/v0.0.64...v0.0.65) (2024-04-09)
+
+### [0.0.64](https://github.com/wrappid/service-core/compare/v0.0.63...v0.0.64) (2024-04-09)
+
+
+### Bug Fixes
+
+* **global:** :bug: post data model return error fix ([36fd208](https://github.com/wrappid/service-core/commit/36fd208d14e411e555261feaf2478a1764d075e9)), closes [#126](https://github.com/wrappid/service-core/issues/126)
+
+### [0.0.63](https://github.com/wrappid/service-core/compare/v0.0.62...v0.0.63) (2024-04-09)
+
+### [0.0.62](https://github.com/wrappid/service-core/compare/v0.0.61...v0.0.62) (2024-04-09)
+
+
+### Features
+
+* **global:** :fire: themeSchema model added in ([f7b86f0](https://github.com/wrappid/service-core/commit/f7b86f09ba6102240bcc5df1c42b855ebe7e6572)), closes [#124](https://github.com/wrappid/service-core/issues/124)
+
+### [0.0.61](https://github.com/wrappid/service-core/compare/v0.0.60...v0.0.61) (2024-04-06)
+
+### [0.0.60](https://github.com/wrappid/service-core/compare/v0.0.59...v0.0.60) (2024-04-06)
+
+
+### Bug Fixes
+
+* **global:** :bug: remove database params from data API ([cf31b30](https://github.com/wrappid/service-core/commit/cf31b3038ba4f65527d890b107aa7cb6507b744f)), closes [#122](https://github.com/wrappid/service-core/issues/122)
+
+### [0.0.59](https://github.com/wrappid/service-core/compare/v0.0.58...v0.0.59) (2024-04-01)
+
+### [0.0.58](https://github.com/wrappid/service-core/compare/v0.0.57...v0.0.58) (2024-04-01)
+
+
+### Features
+
+* **global:** :sparkles: added a new column in Pages Model ([41d16cf](https://github.com/wrappid/service-core/commit/41d16cf1ae4e418473404dd6152914f556b5ce01)), closes [#120](https://github.com/wrappid/service-core/issues/120)
+
+### [0.0.57](https://github.com/wrappid/service-core/compare/v0.0.56...v0.0.57) (2024-04-01)
+
+### [0.0.56](https://github.com/wrappid/service-core/compare/v0.0.55...v0.0.56) (2024-04-01)
+
+
+### Features
+
+* **global:** :sparkles: razorpay integration ([1117760](https://github.com/wrappid/service-core/commit/11177602c0022ec206881ef7e54b3e6f0a9894a2)), closes [#116](https://github.com/wrappid/service-core/issues/116)
+
+
+### Bug Fixes
+
+* **global:** :bug: razorpay payment fix ([46524ec](https://github.com/wrappid/service-core/commit/46524ec06fa6ccdbd1bb2ab3a0160c7c3fcf4893)), closes [#116](https://github.com/wrappid/service-core/issues/116)
+
+### [0.0.55](https://github.com/wrappid/service-core/compare/v0.0.54...v0.0.55) (2024-03-30)
+
+### [0.0.54](https://github.com/wrappid/service-core/compare/v0.0.53...v0.0.54) (2024-03-30)
+
+
+### Features
+
+* **global:** :sparkles: masterdata api moved to service-core ([48a7749](https://github.com/wrappid/service-core/commit/48a7749934b81afeb5a1f0ab6ec823d1fcce454b)), closes [#117](https://github.com/wrappid/service-core/issues/117)
+
 ### [0.0.53](https://github.com/wrappid/service-core/compare/v0.0.52...v0.0.53) (2024-03-21)